+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Use selectedQuantity from state of PDP.
+
 ## [1.0.1] - 2019-06-03
 
 ### Fixed